# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RaceHunter API",
    "version": "0.4.0",
    "description": "Phase 4 finding evidence and bounded Verify Fix replay contract.",
    "license": { "name": "Apache 2.0", "identifier": "Apache-2.0" }
  },
  "servers": [
    {
      "url": "https://{apiHost}",
      "description": "Configured RaceHunter API environment",
      "variables": { "apiHost": { "default": "racehunter-api.internal" } }
    }
  ],
  "security": [],
  "tags": [
    { "name": "Runs", "description": "Durable campaign status" },
    { "name": "Findings", "description": "Verified evidence and immutable replay" }
  ],
  "paths": {
    "/api/runs/{runId}": {
      "get": {
        "tags": ["Runs"],
        "summary": "Read durable campaign status and its verified finding link",
        "operationId": "getRun",
        "parameters": [
          { "$ref": "#/components/parameters/RunId" }
        ],
        "responses": {
          "200": {
            "description": "Durable run state",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Run" } } }
          },
          "404": { "$ref": "#/components/responses/RunNotFound" }
        }
      }
    },
    "/api/findings/{findingId}": {
      "get": {
        "tags": ["Findings"],
        "summary": "Read a verified finding and its immutable replay evidence",
        "operationId": "getFinding",
        "parameters": [
          { "$ref": "#/components/parameters/FindingId" }
        ],
        "responses": {
          "200": {
            "description": "Verified finding projection",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Finding" } } }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/findings/{findingId}/replays": {
      "post": {
        "tags": ["Findings"],
        "summary": "Execute the finding's one server-owned fixed-target verification",
        "description": "Every request supplies an idempotency key, but a finding receives at most one fixed-target execution. Concurrent or later requests return the stored winning attempt without mutating the artifact.",
        "operationId": "verifyFindingFix",
        "parameters": [
          { "$ref": "#/components/parameters/FindingId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyFixRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Fixed replay completed or an already completed result was returned",
            "headers": {
              "Location": { "schema": { "type": "string" }, "description": "Finding resource containing the persisted replay attempt" }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReplayComparison" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "FindingId": {
        "name": "findingId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "RunId": {
        "name": "runId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid replay request",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
      "NotFound": {
        "description": "Finding not found",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
      "Unavailable": {
        "description": "The private replay worker is temporarily unavailable",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
      "RunNotFound": {
        "description": "Run not found"
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string" }
        }
      },
      "VerifyFixRequest": {
        "type": "object",
        "required": ["idempotencyKey"],
        "properties": { "idempotencyKey": { "type": "string", "minLength": 1, "maxLength": 160 } }
      },
      "Run": {
        "type": "object",
        "required": ["id", "status", "maxActors", "maxConcurrentActors", "maxRequests", "maxModelCalls", "maxDurationSeconds", "createdAtUtc"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["Queued", "Running", "Completed", "Failed", "Cancelled"] },
          "maxActors": { "type": "integer", "minimum": 1 },
          "maxConcurrentActors": { "type": "integer", "minimum": 1 },
          "maxRequests": { "type": "integer", "minimum": 1 },
          "maxModelCalls": { "type": "integer", "minimum": 0 },
          "maxDurationSeconds": { "type": "integer", "minimum": 1 },
          "createdAtUtc": { "type": "string", "format": "date-time" },
          "startedAtUtc": { "type": ["string", "null"], "format": "date-time" },
          "completedAtUtc": { "type": ["string", "null"], "format": "date-time" },
          "cancellationRequestedAtUtc": { "type": ["string", "null"], "format": "date-time" },
          "findingId": { "type": ["string", "null"], "format": "uuid", "description": "Verified finding link, populated only after Phase 4 finalization." }
        }
      },
      "ReplayComparison": {
        "type": "object",
        "required": ["vulnerableOutcome", "fixedOutcome", "artifactFingerprint", "idempotencyKey"],
        "properties": {
          "vulnerableOutcome": { "type": "string", "enum": ["Fail"] },
          "fixedOutcome": { "type": "string", "enum": ["Pass", "Fail", "Inconclusive"] },
          "artifactFingerprint": { "type": "string", "pattern": "^sha256:[0-9a-f]{64}$" },
          "idempotencyKey": { "type": "string" }
        }
      },
      "Finding": {
        "type": "object",
        "required": ["id", "runId", "successMessage", "invariantOutcome", "invariantSummary", "traceReferences", "agentInterpretation", "reproductions", "replayArtifact", "timeline", "agentActivity", "replayAttempts"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "runId": { "type": "string", "format": "uuid" },
          "successMessage": { "type": "string" },
          "invariantOutcome": { "type": "string", "enum": ["Fail"] },
          "invariantSummary": { "type": "string" },
          "traceReferences": { "type": "array", "items": { "type": "string" } },
          "agentInterpretation": { "type": "string" },
          "reproductions": { "type": "array", "minItems": 3, "maxItems": 3, "items": { "$ref": "#/components/schemas/Reproduction" } },
          "replayArtifact": { "$ref": "#/components/schemas/ReplayArtifact" },
          "timeline": { "type": "array", "items": { "$ref": "#/components/schemas/ActorLane" } },
          "agentActivity": { "type": "array", "items": { "type": "object" } },
          "replayAttempts": { "type": "array", "items": { "type": "object" } }
        }
      },
      "Reproduction": {
        "type": "object",
        "required": ["attempt", "outcome", "traceReferences"],
        "properties": {
          "attempt": { "type": "integer", "minimum": 1, "maximum": 3 },
          "outcome": { "type": "string", "enum": ["Fail"] },
          "traceReferences": { "type": "array", "items": { "type": "string" } }
        }
      },
      "ReplayArtifact": {
        "type": "object",
        "required": ["id", "fingerprint", "strategy", "seed", "actorCount", "stepCount", "steps"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "fingerprint": { "type": "string", "pattern": "^sha256:[0-9a-f]{64}$" },
          "strategy": { "type": "string" },
          "seed": { "type": "integer" },
          "actorCount": { "type": "integer", "minimum": 2 },
          "stepCount": { "type": "integer", "minimum": 2 },
          "steps": { "type": "array", "items": { "type": "object" } }
        }
      },
      "ActorLane": {
        "type": "object",
        "required": ["actorId", "events"],
        "properties": {
          "actorId": { "type": "integer", "minimum": 1 },
          "events": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["sequence", "attemptId", "stepId", "kind", "requestId", "occurredAtUtc"],
              "properties": {
                "sequence": { "type": "integer" },
                "attemptId": { "type": "string", "format": "uuid" },
                "stepId": { "type": "string" },
                "kind": { "type": "string" },
                "requestId": { "type": "string" },
                "occurredAtUtc": { "type": "string", "format": "date-time" }
              }
            }
          }
        }
      }
    }
  }
}